% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{assign_interview__keys}
\alias{assign_interview__keys}
\title{Function to assign interview key to spatial data}
\usage{
assign_interview__keys(c, e)
}
\arguments{
\item{Connection}{to postgres}

\item{imput}{ed}
}
\value{
the completed query and checks if there are errors.
}
\description{
Function to assign interview key to spatial data
}
