% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{autonumber_ed}
\alias{autonumber_ed}
\title{Function to AutoNumber buildings in the spatial database.}
\usage{
autonumber_ed(c, ed_list)
}
\arguments{
\item{c}{PostgreSQL Connection.}

\item{ed_list}{Single ED or a list of ED.}
}
\value{
Returns the completed query.
}
\description{
Function to AutoNumber buildings in the spatial database.
}
