% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{clear_ed}
\alias{clear_ed}
\title{Function to clear some ed data for specific columns from PostgreSQL.}
\usage{
clear_ed(c, e, t)
}
\arguments{
\item{e}{imput ed.}

\item{t}{imput table name to clear.}

\item{C}{connection to PostgreSQL Database.}
}
\value{
Completed is returned if successful.
}
\description{
Function to clear some ed data for specific columns from PostgreSQL.
}
