% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{get_completed_ed}
\alias{get_completed_ed}
\title{Function to get the amount of completed eds and see which ones need to be added to mics7_dp_assignment table.}
\usage{
get_completed_ed()
}
\value{
Returns the completed query, and a list of counts.
}
\description{
Function to get the amount of completed eds and see which ones need to be added to mics7_dp_assignment table.
}
