% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{get_vr}
\alias{get_vr}
\title{Function to get visitation records and compare it to the spatail data.}
\usage{
get_vr(ed_list, location)
}
\arguments{
\item{imput}{ed}

\item{input}{location}
}
\value{
the completed query, and a list of counts.
}
\description{
Function to get visitation records and compare it to the spatail data.
}
