% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{reconcile_vr_spatial}
\alias{reconcile_vr_spatial}
\title{Function to Reconcile non spatial data to spatial database.}
\usage{
reconcile_vr_spatial(c, e)
}
\arguments{
\item{c}{PostgreSQL Connection.}

\item{e}{Single ED or a list of ED.}
}
\value{
Returns the completed query.
}
\description{
Function to Reconcile non spatial data to spatial database.
}
