% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{sequence_check}
\alias{sequence_check}
\title{Function to sort spatial blocks and buildings and check if they are in sequence.}
\usage{
sequence_check(ed_list, location)
}
\arguments{
\item{location}{input location}

\item{ed}{imput ed. Make it a list if you have multiple ed.}
}
\value{
Returns the completed query, and a list of counts.
}
\description{
Function to sort spatial blocks and buildings and check if they are in sequence.
}
