% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{update_interview__id}
\alias{update_interview__id}
\title{Function to Update interivew__id from MySQL table to PostgreSQL Table.}
\usage{
update_interview__id(c, e)
}
\arguments{
\item{c}{PostgreSQL Connection.}

\item{e}{Single ED or a list of ED.}
}
\value{
Returns the completed query.
}
\description{
Function to Update interivew__id from MySQL table to PostgreSQL Table.
}
