% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{clear_ed}
\alias{clear_ed}
\title{Function to clear data}
\usage{
clear_ed(c, e, t)
}
\arguments{
\item{Connection}{to postgres}

\item{imput}{ed}
}
\value{
the completed query.
}
\description{
Function to clear data
}
