% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.r
\name{get_vr}
\alias{get_vr}
\title{Function to get visitation records in csv and compare it to the spatail data.}
\usage{
get_vr(ed_list, location)
}
\arguments{
\item{location}{Input location for where the files will be stored.}

\item{imput}{ed in a list.}
}
\value{
Returns a csv, the completed query, and a list of counts.
}
\description{
Function to get visitation records in csv and compare it to the spatail data.
}
